feat: move Lookup Classes to Classes
