--- components/class/ClassOverview.tsx
@@ -1,201 +1,0 @@
-// External libraries
-import { useSupabaseClient } from "@supabase/auth-helpers-react";
-import Link from "next/link";
-import { useTranslation } from "next-i18next";
-import { FC, useContext } from "react";
-
-// SK Components
-import {
-  Card,
-  CardHeader,
-  Columns,
-  ContentLayout,
-  Header,
-  Section,
-  Snackbar,
-} from "@suankularb-components/react";
-
-// Internal components
-import ContactsSection from "@/components/account/ContactSection";
-import DynamicAvatar from "@/components/common/DynamicAvatar";
-
-// Contexts
-import SnackbarContext from "@/contexts/SnackbarContext";
-
-// Backend
-// import {
-//   addContactToClassroom,
-//   removeContactFromClassroom,
-// } from "@/utils/backend/classroom/classroom";
-// import {
-//   createContact,
-//   deleteContact,
-//   updateContact,
-// } from "@/utils/backend/contact";
-
-// Helpers
-import getLocaleName from "@/utils/helpers/getLocaleName";
-import getLocaleString from "@/utils/helpers/getLocaleString";
-
-// Hooks
-import useLocale from "@/utils/helpers/useLocale";
-import useRefreshProps from "@/utils/helpers/useRefreshProps";
-import { Classroom } from "@/utils/types/classroom";
-import { Contact } from "@/utils/types/contact";
-import { Teacher } from "@/utils/types/person";
-import createContact from "@/utils/backend/contact/createContact";
-import updateContact from "@/utils/backend/contact/updateContact";
-
-// Types
-// import { ClassOverview as ClassOverviewType } from "@/utils/types/class";
-// import { Contact } from "@/utils/types/contact";
-// import { Teacher } from "@/utils/types/person";
-
-/**
- * Displays a list of this class’ Class Advisors.
- *
- * @param advisors The list of Advisors to display.
- *
- * @returns A Section.
- */
-const AdvisorsSection: FC<{ advisors: Classroom["class_advisors"] }> = ({
-  advisors,
-}) => {
-  const locale = useLocale();
-  const { t } = useTranslation("class");
-
-  return (
-    <Section>
-      <Header>{t("overview.classAdvisors.title")}</Header>
-      <Columns columns={3}>
-        {advisors.map((teacher) => (
-          <Card
-            key={teacher.id}
-            appearance="outlined"
-            stateLayerEffect
-            href={`/lookup/person?id=${teacher.id}&role=teacher`}
-            element={Link}
-          >
-            <CardHeader
-              avatar={<DynamicAvatar profile={teacher.profile} />}
-              title={getLocaleName(locale, teacher)}
-              subtitle={getLocaleString(teacher.subject_group.name, locale)}
-            />
-          </Card>
-        ))}
-      </Columns>
-    </Section>
-  );
-};
-
-/**
- * Displays a list of this class’ Contacts. The Class Advisors of this class
- * can also edit this list.
- *
- * @param advisors The list of Contacts to display.
- */
-const ClassContactsSection: FC<{
-  // contacts: Contact[];
-  contacts: Contact[];
-  classID?: string;
-  editable?: boolean;
-}> = ({ contacts, classID, editable }) => {
-  const { t } = useTranslation("common");
-  const { setSnackbar } = useContext(SnackbarContext);
-  const refreshProps = useRefreshProps();
-  const supabase = useSupabaseClient();
-
-  /**
-   * Creates and add a Contact to this class.
-   *
-   * @param contact The Contact to add.
-   */
-  async function handleAdd(contact: Contact) {
-    const { data: contactID, error } = await createContact(supabase, contact);
-    if (error) {
-      setSnackbar(<Snackbar>{t("snackbar.failure")}</Snackbar>);
-      return;
-    }
-
-    const { error: classError } = await supabase.from("classroom_contacts").insert({
-      classroom_id: classID,
-      contact_id: contactID,
-    });
-
-    if (classError) {
-      setSnackbar(<Snackbar>{t("snackbar.failure")}</Snackbar>);
-      return;
-    }
-
-    refreshProps();
-  }
-
-  /**
-   * Edits a given Contact.
-   *
-   * @param contact The new data for the Contact.
-   */
-  async function handleEdit(contact: Contact) {
-    const { error } = await updateContact(supabase, contact);
-    if (error) {
-      setSnackbar(<Snackbar>{t("snackbar.failure")}</Snackbar>);
-      return;
-    }
-
-    refreshProps();
-  }
-
-  /**
-   * Delete a given Contact from Supabase.
-   *
-   * @param contactID The ID of the Contact to delete.
-   */
-  async function handleRemove(contactID: string) {
-    const { error } = await supabase.from("contacts").delete().match({ id: contactID });
-    if (error) {
-      setSnackbar(<Snackbar>{t("snackbar.failure")}</Snackbar>);
-      return;
-    }
-
-    refreshProps();
-  }
-
-  return (
-    <ContactsSection
-      contacts={contacts}
-      handleAdd={editable ? handleAdd : undefined}
-      handleEdit={editable ? handleEdit : undefined}
-      handleRemove={
-        editable ? (contactID) => handleRemove(contactID) : undefined
-      }
-    />
-  );
-};
-
-/**
- * Displays an overview of a Class. Used for Class Overview pages like `/class`
- * and `/lookup/class/[classNumber]`. Occupies the full page.
- *
- * @param classroom The overview information (Advisors and Contacts) of the Class to display.
- * @param editable If the user can edit information about this class. Reserved for this class’ Class Advisors and admins.
- *
- * @returns A Content Layout.
- */
-const ClassOverview: FC<{
-  // classroom: ClassOverviewType;
-  classroom: Pick<Classroom, "id" | "class_advisors" | "contacts">;
-  editable?: boolean;
-}> = ({ classroom, editable }) => {
-  return (
-    <ContentLayout>
-      <AdvisorsSection advisors={classroom.class_advisors} />
-      <ClassContactsSection
-        contacts={classroom.contacts}
-        classID={classroom.id}
-        editable={editable}
-      />
-    </ContentLayout>
-  );
-};
-
-export default ClassOverview;
