--- public/static/locales/en-US/class.json
@@ -1,65 +1,0 @@
-{
-  "overview": {
-    "title": {
-      "student": "Your class",
-      "teacher": "Your class",
-      "lookup": "M.{{number}}"
-    },
-    "classAdvisors": {
-      "title": "Class advisors"
-    }
-  },
-  "student": {
-    "title": {
-      "student": "Your classmates",
-      "teacher": "Students in your class",
-      "lookup": "Students of M.{{number}}"
-    },
-    "list": {
-      "searchAlt": "Search students",
-      "action": {
-        "print": "Print",
-        "saveVCards": "Save to contacts"
-      }
-    },
-    "print": {
-      "title": "Print student list",
-      "language": "Language",
-      "columns": {
-        "label": "Information to show",
-        "classNo": "Class №",
-        "studentID": "Student ID",
-        "prefix": "Prefix",
-        "fullName": "Full name",
-        "nickname": "Nickname",
-        "allergies": "Allergies (in Thai)",
-        "shirtSize": "Shirt size",
-        "pantsSize": "School pants size"
-      },
-      "numEmpty": "Empty columns",
-      "enableNotes": "Notes column",
-      "enableTimestamp": "Timestamp"
-    }
-  },
-  "teacher": {
-    "title": {
-      "student": "Your teachers",
-      "teacher": "Teaching this class",
-      "lookup": "Teachers of M.{{number}}"
-    },
-    "list": {
-      "searchAlt": "Search teachers"
-    }
-  },
-  "schedule": {
-    "title": "M.{{number}} schedule"
-  },
-  "common": {
-    "navigation": {
-      "overview": "Overview",
-      "students": "Students",
-      "teachers": "Teachers",
-      "schedule": "Schedule"
-    }
-  }
-}
