+++ public/static/locales/en-US/lookup.json
@@ -10,10 +10,6 @@
       "title": "Lookup <strong>teachers</strong>",
       "desc": "Find a teacher’s subjects, office, schedule, contacts, and more."
     },
-    "classes": {
-      "title": "Lookup <strong>classes</strong>",
-      "desc": "See a list of a class’ students and teachers, get a printout, and more."
-    },
     "documents": {
       "title": "Lookup <strong>documents</strong>",
       "desc": "Easily locate, read, and download school orders and documents."
