--- public/static/locales/th/class.json
@@ -1,65 +1,0 @@
-{
-  "overview": {
-    "title": {
-      "student": "ห้องเรียน",
-      "teacher": "ห้องที่ปรึกษา",
-      "lookup": "ม.{{number}}"
-    },
-    "classAdvisors": {
-      "title": "ครูที่ปรึกษา"
-    }
-  },
-  "student": {
-    "title": {
-      "student": "นักเรียนในห้อง",
-      "teacher": "นักเรียนในห้อง",
-      "lookup": "นักเรียนห้องม.{{number}}"
-    },
-    "list": {
-      "searchAlt": "ค้นหานักเรียน",
-      "action": {
-        "print": "พิมพ์",
-        "saveVCards": "บันทึกลงแอพรายชื่อ"
-      }
-    },
-    "print": {
-      "title": "พิมพ์ใบรายชื่อ",
-      "language": "ภาษา",
-      "columns": {
-        "label": "ข้อมูลที่แสดง",
-        "classNo": "เลขที่ในห้อง",
-        "studentID": "เลขประจำตัวนักเรียน",
-        "prefix": "คำนำหน้าชื่อ",
-        "fullName": "ชื่อจริง-นามสกุล",
-        "nickname": "ชื่อเล่น",
-        "allergies": "ภูมิแพ้/อาหารที่แฟ้ (ภาษาไทย)",
-        "shirtSize": "ไซส์เสื้อ",
-        "pantsSize": "ขนาดกางเกงนักเรียน"
-      },
-      "numEmpty": "เพิ่มหลัก",
-      "enableNotes": "หลักหมายเหตุ",
-      "enableTimestamp": "ประทับเวลา"
-    }
-  },
-  "teacher": {
-    "title": {
-      "student": "อาจารย์ที่สอนห้องนี้",
-      "teacher": "อาจารย์ที่สอนห้องนี้",
-      "lookup": "อาจารย์ที่สอนม.{{number}}"
-    },
-    "list": {
-      "searchAlt": "ค้นหาอาจารย์"
-    }
-  },
-  "schedule": {
-    "title": "ตารางเรียนม.{{number}}"
-  },
-  "common": {
-    "navigation": {
-      "overview": "ข้อมูล",
-      "students": "นักเรียน",
-      "teachers": "อาจารย์",
-      "schedule": "ตารางเรียน"
-    }
-  }
-}
