+++ public/static/locales/th/classes.json
@@ -1,0 +1,93 @@
+{
+  "title": "ห้องเรียน",
+  "list": {
+    "yourClass": "ห้องเรียนของคุณ",
+    "item": {
+      "period": {
+        "upcoming": "คาบต่อไปใน {{duration}}",
+        "finished": "เลิกเรียนแล้ว"
+      }
+    }
+  },
+  "header": {
+    "action": {
+      "print": "พิมพ์",
+      "saveContact": "บันทึก",
+      "attendance": {
+        "view": "ข้อมูลเช็คชื่อ",
+        "take": "เช็คชื่อ"
+      },
+      "requestInfo": "ขอข้อมูลเพิ่มเติม"
+    }
+  },
+  "detail": {
+    "general": {
+      "classAdvisors": "ครูที่ปรึกษา",
+      "room": "ห้องที่ใช้เรียน"
+    },
+    "contacts": {
+      "title": "ช่องทางการติดต่อ"
+    }
+  },
+  "dialog": {
+    "attendance": {
+      "title": {
+        "view": "ข้อมูลเช็คชื่อ",
+        "take": "เช็คชื่อ"
+      },
+      "event": {
+        "assembly": {
+          "name": "เข้าแถว",
+          "desc": "คุณสามารถเช็คชื่อเข้าแถวได้<strong>ก่อนเวลา 08:00 น.</strong> เท่านั้น คุณจะ<strong>ไม่สามารถ</strong>บันทึกข้อมูลการเช็คชื่อได้หลังจากเวลานี้"
+        },
+        "homeroom": {
+          "name": "โฮมรูม",
+          "desc": "คุณสามารถเช็คชื่อโฮมรูมไก้<strong>ระหว่างเวลา 08:00–08:30 น.</strong> คุณจะ<strong>ไม่สามารถ</strong>บันทึกข้อมูลการเช็คชื่อได้หลังจากเวลานี้"
+        }
+      },
+      "item": {
+        "classNo": "เลขที่ {{classNo}}",
+        "presence": {
+          "present": "มา",
+          "absent": "ขาด"
+        },
+        "absenceType": {
+          "sick": "ลาป่วย",
+          "business": "ลากิจ",
+          "activity": "กิจกรรมโรงเรียน",
+          "other": "อื่น ๆ"
+        },
+        "enterReason": "ระบุเหตุผล"
+      },
+      "action": {
+        "save": "บันทึก"
+      }
+    },
+    "confirmAttendanceSave": {
+      "title": "ยืนยันการบันทึก",
+      "desc": "เมื่อคุณบันทึกข้อมูลการเช็คชื่อแล้ว คุณจะไม่สามารถแก้ไขได้ในภายหลัง",
+      "action": {
+        "goBack": "แก้ไขต่อ",
+        "confirm": "ยืนยัน"
+      }
+    }
+  },
+  "print": {
+    "title": "พิมพ์ใบรายชื่อ",
+    "language": "ภาษา",
+    "columns": {
+      "label": "ข้อมูลที่แสดง",
+      "classNo": "เลขที่ในห้อง",
+      "studentID": "เลขประจำตัวนักเรียน",
+      "prefix": "คำนำหน้าชื่อ",
+      "fullName": "ชื่อจริง-นามสกุล",
+      "nickname": "ชื่อเล่น",
+      "allergies": "ภูมิแพ้/อาหารที่แฟ้ (ภาษาไทย)",
+      "shirtSize": "ไซส์เสื้อ",
+      "pantsSize": "ขนาดกางเกงนักเรียน"
+    },
+    "numEmpty": "เพิ่มหลัก",
+    "enableNotes": "หลักหมายเหตุ",
+    "enableTimestamp": "ประทับเวลา"
+  }
+}
