+++ public/static/locales/th/lookup.json
@@ -10,10 +10,6 @@
       "title": "ค้นหา<strong>อาจารย์</strong>",
       "desc": "ค้นหาข้อมูลห้องที่ปรึกษา ตารางสอน ช่องทางติดต่อ และอื่น ๆ ของครูในโรงเรียน"
     },
-    "classes": {
-      "title": "ค้นหา<strong>ห้องเรียน</strong>",
-      "desc": "ดูรายชื่อนักเรียน อาจารย์ วิชา และตารางเรียนของทุกห้องในโรงเรียน"
-    },
     "documents": {
       "title": "ค้นหา<strong>เอกสาร</strong>",
       "desc": "ค้นหาคำสั่งโรงเรียนและหนังสือออกของโรงเรียนอย่างง่ายดาย"
